mentor_review_files_2026-05-08/requirements.txt
torch>=2.1
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
tqdm>=4.66
matplotlib>=3.8
tables>=3.11
